[fix] 보관함-내가그린코스 다중삭제 구현
selection 라이브러리를 안 쓰고 지훈이가 쓴 방법을 따라주었음. 이 페이지에는 '삭제' 말고도 여러 기능들이 붙어있는데 그것들은 아직 안 고쳐졌고 수정을 해주어야 함.

+++ app/src/main/java/com/runnect/runnect/presentation/mydrawdetail/MyDrawDetailActivity.kt
@@ -28,7 +28,7 @@ class MyDrawDetailActivity :
 
     val viewModel: MyDrawDetailViewModel by viewModels()
 
-    val selectList = arrayListOf<Long>()
+    val selectList = arrayListOf<Int>()
 
     lateinit var departureLatLng: LatLng
     private val touchList = arrayListOf<LatLng>()
@@ -92,7 +92,7 @@ class MyDrawDetailActivity :
         val courseId = intent.getIntExtra("fromStorageFragment", 0)
         Timber.tag(ContentValues.TAG).d("courseId from Storage : $courseId")
 
-        selectList.add(courseId.toLong()) //courseId를 지역변수로 선언해서 여기에 작성해주었음
+        selectList.add(courseId) //courseId를 지역변수로 선언해서 여기에 작성해주었음
         viewModel.getMyDrawDetail(courseId = courseId)
     }
